User_Service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["Muhammad Junaid <muhammadjunaid7015140@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.31.0"}
sqlmodel = "^0.0.22"
python-dotenv = "^1.0.1"
mypy = "^1.11.2"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.12"
passlib = "^1.7.4"


[tool.poetry.scripts]
dev= "user_service.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
